Move VariantGroupedtaskStatusBadge to folder it is used in
